# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Sorolens API
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  description: |
    Sorolens indexes and surfaces data from Soroban smart contracts: events,
    invocations, storage, and per-contract analytics. This spec is the single
    source of truth for the auto-generated Go client (packages/go-client) and
    matches the live routes in apps/api/internal/router/router.go.

    Authentication: most read endpoints accept an optional API key via
    `Authorization: Bearer <token>` or `X-API-Key: <token>`. API-key management
    additionally requires an admin identity via role headers
    (`X-User-ID`, `X-GitHub-Login`, `X-User-Sub`, or `X-GitHub-ID`).
  contact:
    name: Sorolens
servers:
  - url: https://api.sorolens.xyz
    description: Production
...
paths:
...
  /api/v1/stats/global:
    get:
      tags: [Stats]
      summary: Global index statistics
      operationId: getGlobalStats
      security:
        - ApiKeyAuth: [read:contracts]
      responses:
        "200":
          description: Global counts across all tracked contracts
          content:
            application/json:
              schema:
                type: object
                required: [tracked_contracts, total_events, total_invocations, total_storage_entries]
                properties:
                  tracked_contracts:
                    type: integer
                    format: int64
                  total_events:
                    type: integer
                    format: int64
                  total_invocations:
                    type: integer
                    format: int64
                  total_storage_entries:
                    type: integer
                    format: int64
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/invocations:
    get:
      tags: [Invocations]
      summary: List invocations across all tracked contracts
      description: |
        Global invocation explorer. Results are ordered newest first
        (ledger DESC, tx_hash DESC) and paginated with an opaque cursor that
        encodes the (ledger, tx_hash) keyset position.
      operationId: listAllInvocations
      security:
        - ApiKeyAuth: [read:contracts]
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/NetworkParam"
        - name: contract_id
          in: query
          required: false
          schema:
            type: string
          description: Restrict results to a single contract ID.
        - name: fn
          in: query
          required: false
          schema:
            type: string
          description: Function name filter.
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [SUCCESS, FAILED, NOT_FOUND]
          description: Invocation status filter.
        - name: from
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
          description: Lower ledger bound (inclusive).
        - name: to
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
          description: Upper ledger bound (inclusive).
        - name: since
          in: query
          required: false
          schema:
            type: string
          description: Lower bound on ledger_closed_at (YYYY-MM-DD or RFC3339, inclusive).
        - name: until
          in: query
          required: false
          schema:
            type: string
          description: Upper bound on ledger_closed_at (YYYY-MM-DD or RFC3339, inclusive).
      responses:
        "200":
          description: Page of invocations across every tracked contract
          content:
            application/json:
              schema:
                type: object
                required: [invocations, next_cursor]
                properties:
                  invocations:
                    type: array
                    items:
                      $ref: "#/components/schemas/Invocation"
                  next_cursor:
                    type: string
        "422":
          $ref: "#/components/responses/InvalidInput"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/contracts:
    get:
      tags: [Contracts]
      summary: List tracked contracts
      operationId: listContracts
      security:
        - ApiKeyAuth: [read:contracts]
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/NetworkParam"
        - name: status
          in: query
          required: false
          schema:
            type: string
          description: Filter by contract status (pending, active, backfilling, error, paused).
      responses:
        "200":
          description: Page of contracts
          content:
            application/json:
              schema:
                type: object
                required: [contracts, next_cursor]
                properties:
                  contracts:
                    type: array
                    items:
                      $ref: "#/components/schemas/Contract"
                  next_cursor:
                    type: string
                    description: Opaque cursor for the next page; empty when no more pages.
        "422":
          $ref: "#/components/responses/InvalidInput"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [Contracts]
      summary: Register a contract for tracking
      operationId: registerContract
      security:
        - ApiKeyAuth: [write:contracts]
        - RoleAuth: [contributor]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id, network]
              properties:
                id:
                  type: string
                  pattern: "^C[A-Za-z0-9]{55}$"
                  description: Stellar contract ID (56 chars, starts with C).
                network:
                  type: string
                  enum: [testnet, mainnet, futurenet, standalone]
                label:
                  type: string
                  description: Optional human-friendly label.
      responses:
        "201":
          description: Contract registered (status pending until backfill completes).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/ForbiddenRole"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "422":
          $ref: "#/components/responses/InvalidInput"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/contracts/{id}:
    get:
      tags: [Contracts]
      summary: Get one contract
      operationId: getContract
      security:
        - ApiKeyAuth: [read:contracts]
      parameters:
        - $ref: "#/components/parameters/ContractID"
      responses:
        "200":
          description: The contract
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/contracts/{id}/events:
    get:
      tags: [Events]
      summary: List contract events
      operationId: listContractEvents
      security:
        - ApiKeyAuth: [read:contracts]
      parameters:
        - $ref: "#/components/parameters/ContractID"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/NetworkParam"
        - name: type
          in: query
          required: false
          schema:
            type: string
          description: Event type filter.
        - name: from
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
          description: Lower ledger bound (inclusive).
        - name: to
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
          description: Upper ledger bound (inclusive).
      responses:
        "200":
          description: Page of events
          content:
            application/json:
              schema:
                type: object
                required: [events, next_cursor]
                properties:
                  events:
                    type: array
                    items:
                      $ref: "#/components/schemas/Event"
                  next_cursor:
                    type: string
        "422":
          $ref: "#/components/responses/InvalidInput"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/contracts/{id}/invocations:
    get:
      tags: [Invocations]
      summary: List contract invocations
      operationId: listContractInvocations
      security:
        - ApiKeyAuth: [read:contracts]
      parameters:
        - $ref: "#/components/parameters/ContractID"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/NetworkParam"
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [SUCCESS, FAILED, NOT_FOUND]
          description: Invocation status filter.
        - name: fn
          in: query
          required: false
          schema:
            type: string
          description: Function name filter.
        - name: from
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
          description: Lower ledger bound (inclusive).
        - name: to
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
          description: Upper ledger bound (inclusive).
      responses:
        "200":
          description: Page of invocations
          content:
            application/json:
              schema:
                type: object
                required: [invocations, next_cursor]
                properties:
                  invocations:
                    type: array
                    items:
                      $ref: "#/components/schemas/Invocation"
                  next_cursor:
                    type: string
        "422":
          $ref: "#/components/responses/InvalidInput"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/contracts/{id}/storage:
    get:
      tags: [Storage]
      summary: List contract storage entries
      operationId: listContractStorage
      security:
        - ApiKeyAuth: [read:contracts]
      parameters:
        - $ref: "#/components/parameters/ContractID"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/NetworkParam"
        - name: durability
          in: query
          required: false
          schema:
            type: string
            enum: [temporary, persistent, instance]
          description: Storage durability filter.
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [live, archived, deleted]
          description: Storage entry status filter.
      responses:
        "200":
          description: Page of storage entries
          content:
            application/json:
              schema:
                type: object
                required: [storage, next_cursor]
                properties:
                  storage:
                    type: array
                    items:
                      $ref: "#/components/schemas/StorageEntry"
                  next_cursor:
                    type: string
        "422":
          $ref: "#/components/responses/InvalidInput"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/contracts/{id}/stats:
    get:
      tags: [Stats]
      summary: Per-contract statistics
      operationId: getContractStats
      security:
        - ApiKeyAuth: [read:contracts]
      parameters:
        - $ref: "#/components/parameters/ContractID"
        - name: window
          in: query
          required: false
          schema:
            type: string
          description: Aggregation window. 7d and 30d map to 7/30 days; anything else is 24 hours.
      responses:
        "200":
          description: Contract statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractStats"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/contracts/{id}/forecast:
    get:
      tags: [Forecast]
      summary: Forecast contract usage over the next N days
      operationId: getContractForecast
      security:
        - ApiKeyAuth: [read:contracts]
      parameters:
        - $ref: "#/components/parameters/ContractID"
        - name: horizon
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 365
          description: Number of days to forecast (default 30; optional trailing d accepted).
      responses:
        "200":
          description: Forecast series for fees, invocations, and events
          content:
            application/json:
              schema:
                type: object
                required: [contract_id, lookback_days, series]
                properties:
                  contract_id:
                    type: string
                  lookback_days:
                    type: integer
                  series:
                    type: array
                    items:
                      $ref: "#/components/schemas/ForecastSeries"
        "400":
          description: Invalid horizon parameter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/contracts/{id}/snapshot:
    get:
      tags: [Snapshots]
      summary: Snapshot of contract storage at a historical ledger
      operationId: getContractSnapshot
      security:
        - ApiKeyAuth: [read:contracts]
      parameters:
        - $ref: "#/components/parameters/ContractID"
        - name: ledger
          in: query
          required: true
          schema:
            type: integer
            format: int32
            minimum: 1
          description: Ledger to snap (greater than zero).
      responses:
        "200":
          description: Storage snapshot plus the last event at or before the ledger
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractSnapshot"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/InvalidInput"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/contracts/{id}/stream:
    get:
      tags: [Events]
      summary: Poll the most recent contract events
      operationId: streamContractEvents
      description: |
        Short-poll convenience used by the dashboard. Returns the 20 most
        recent events as JSON (not SSE).
      security:
        - ApiKeyAuth: [read:contracts]
      parameters:
        - $ref: "#/components/parameters/ContractID"
      responses:
        "200":
          description: Most recent events
          headers:
            Last-Modified:
              schema:
                type: string
                format: date-time
          content:
            application/json:
              schema:
                type: object
                required: [events]
                properties:
                  events:
                    type: array
                    items:
                      $ref: "#/components/schemas/Event"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/api-keys:
    get:
      tags: [API Keys]
      summary: List API keys
      operationId: listApiKeys
      security:
        - ApiKeyAuth: [admin:*]
        - RoleAuth: [admin]
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Page of API keys
          content:
            application/json:
              schema:
                type: object
                required: [api_keys, next_cursor]
                properties:
                  api_keys:
                    type: array
                    items:
                      $ref: "#/components/schemas/APIKey"
                  next_cursor:
                    type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/ForbiddenScope"
        "422":
          $ref: "#/components/responses/InvalidInput"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [API Keys]
      summary: Create an API key
      operationId: createApiKey
      security:
        - ApiKeyAuth: [admin:*]
        - RoleAuth: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, scopes]
              properties:
                name:
                  type: string
                  description: Human-friendly key name.
                scopes:
                  type: array
                  items:
                    type: string
                    enum: [read:contracts, write:contracts, read:watchdog, admin:*]
                  minItems: 1
                  description: Scopes granted to the key.
      responses:
        "201":
          description: Key created; the plaintext key is returned once.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/APIKey"
                  - type: object
                    required: [key]
                    properties:
                      key:
                        type: string
                        description: Plaintext token, shown once.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/ForbiddenScope"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "422":
          $ref: "#/components/responses/InvalidInput"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/api-keys/{id}:
    delete:
      tags: [API Keys]
      summary: Revoke an API key
      operationId: revokeApiKey
      security:
        - ApiKeyAuth: [admin:*]
        - RoleAuth: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: API key ID.
      responses:
        "204":
          description: Key revoked
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/ForbiddenScope"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/v1/watchdog/stats:
    get:
      tags: [Watchdog]
      summary: Watchdog statistics
      operationId: getWatchdogStats
      security:
        - ApiKeyAuth: [read:watchdog]
      parameters:
        - $ref: "#/components/parameters/NetworkParam"
      responses:
        "200":
          description: Watchdog summary counts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WatchdogStats"
        "422":
          $ref: "#/components/responses/InvalidInput"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/watchdog/alerts:
    get:
      tags: [Watchdog]
      summary: List watchdog alerts
      operationId: listWatchdogAlerts
      security:
        - ApiKeyAuth: [read:watchdog]
      parameters:
        - name: severity
          in: query
          required: false
          schema:
            type: string
            enum: [Info, Warning, Critical]
          description: Severity filter.
        - $ref: "#/components/parameters/NetworkParam"
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 500
          description: Maximum alerts to return (default 100).
      responses:
        "200":
          description: Alerts
          content:
            application/json:
              schema:
                type: object
                required: [alerts]
                properties:
                  alerts:
                    type: array
                    items:
                      $ref: "#/components/schemas/ContractAlert"
        "422":
          $ref: "#/components/responses/InvalidInput"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/watchdog/contracts:
    get:
      tags: [Watchdog]
      summary: List monitored contracts
      operationId: listMonitoredContracts
      security:
        - ApiKeyAuth: [read:watchdog]
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/NetworkParam"
      responses:
        "200":
          description: Page of monitored contracts
          content:
            application/json:
              schema:
                type: object
                required: [contracts, next_cursor]
                properties:
                  contracts:
                    type: array
                    items:
                      $ref: "#/components/schemas/MonitoredContract"
                  next_cursor:
                    type: string
        "422":
          $ref: "#/components/responses/InvalidInput"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/watchdog/contracts/{id}:
    get:
      tags: [Watchdog]
      summary: Get one monitored contract
      operationId: getMonitoredContract
      security:
        - ApiKeyAuth: [read:watchdog]
      parameters:
        - $ref: "#/components/parameters/ContractID"
      responses:
        "200":
          description: The monitored contract
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonitoredContract"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/v1/watchdog/contracts/{id}/alerts:
    get:
      tags: [Watchdog]
      summary: List alerts for a monitored contract
      operationId: listContractAlerts
      security:
        - ApiKeyAuth: [read:watchdog]
      parameters:
        - $ref: "#/components/parameters/ContractID"
        - name: severity
          in: query
          required: false
          schema:
            type: string
            enum: [Info, Warning, Critical]
        - $ref: "#/components/parameters/NetworkParam"
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 500
      responses:
        "200":
          description: Alerts for the contract
          content:
            application/json:
              schema:
                type: object
                required: [alerts]
                properties:
                  alerts:
                    type: array
                    items:
                      $ref: "#/components/schemas/ContractAlert"
        "422":
          $ref: "#/components/responses/InvalidInput"
        "500":
          $ref: "#/components/responses/InternalError"
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      description: "Scoped API key. Send as `Authorization: Bearer <token>` (or `X-API-Key`)."
    RoleAuth:
      type: apiKey
      in: header
      name: X-User-ID
      description: |
        Role-based auth. Resolve a user identity from one of
        X-User-ID, X-GitHub-Login, X-User-Sub, or X-GitHub-ID and enforce the
        minimum role declared next to this security scheme.
    IdentityHeaderAuth:
      type: apiKey
      in: header
      name: X-User-ID
      description: Watchlist endpoints read the user ID from X-User-ID.
  parameters:
    ContractID:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Contract ID (starts with C).
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
      description: Opaque pagination cursor from a previous response.
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
      description: Page size (default 50, clamped to 200).
    NetworkParam:
      name: network
      in: query
      required: false
      schema:
        type: string
        enum: [testnet, mainnet, futurenet, standalone]
      description: Network filter; empty means all networks.
...
  responses:
    Unauthorized:
      description: Missing or invalid credential or identity
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    ForbiddenRole:
      description: Identity present but role below the required minimum
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/RoleError"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InvalidInput:
      description: Invalid query or body input
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    UnsupportedMediaType:
      description: Content-Type is not application/json
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    RateLimited:
      description: Global per-IP rate limit exceeded
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Contract:
      type: object
      required: [id, network, label, wasm_hash, created_at_ledger, backfill_complete_at, status, added_at]
      properties:
        id:
          type: string
        network:
          type: string
        label:
          type: string
        wasm_hash:
          type: string
        created_at_ledger:
          type: integer
          format: int64
        backfill_complete_at:
          type: string
          format: date-time
          nullable: true
        status:
          type: string
          enum: [pending, active, backfilling, error, paused]
        added_at:
          type: string
          format: date-time
    Event:
      type: object
      required:
        - id
        - contract_id
        - network
        - ledger
        - ledger_closed_at
        - tx_hash
        - type
        - topic_xdr
        - value_xdr
        - topic_decoded
        - value_decoded
        - in_successful_call
      properties:
        id:
          type: string
        contract_id:
          type: string
        network:
          type: string
        ledger:
          type: integer
          format: int32
        ledger_closed_at:
          type: string
          format: date-time
        tx_hash:
          type: string
        type:
          type: string
        topic_xdr:
          type: array
          items:
            type: string
        value_xdr:
          type: string
        topic_decoded:
          type: array
          items: {}
        value_decoded: {}
        in_successful_call:
          type: boolean
    Invocation:
      type: object
      required:
        - tx_hash
        - contract_id
        - network
        - ledger
        - ledger_closed_at
        - status
        - function_name
        - args_decoded
        - result_decoded
        - result_xdr
        - resource_fee_charged
        - cpu_insn
        - mem_byte
        - ledger_read_byte
        - ledger_write_byte
        - application_order
      properties:
        tx_hash:
          type: string
        contract_id:
          type: string
        network:
          type: string
        ledger:
          type: integer
          format: int32
        ledger_closed_at:
          type: string
          format: date-time
        status:
          type: string
        function_name:
          type: string
        args_decoded:
          type: object
          additionalProperties: true
        result_decoded: {}
        result_xdr:
          type: string
        resource_fee_charged:
          type: integer
          format: int64
        cpu_insn:
          type: integer
          format: int64
        mem_byte:
          type: integer
          format: int64
        ledger_read_byte:
          type: integer
          format: int64
        ledger_write_byte:
          type: integer
          format: int64
        application_order:
          type: integer
...